modify and and folders

+++ authentification/otpAuthConnect.tsx
@@ -4,81 +4,34 @@ import Back from '@/components/btnBack';
 import { app, auth, firestore, PhoneAuthProvider, signInWithCredential } from '@/firebase.config';
 import { OtpInput } from "react-native-otp-entry";
 import { FirebaseRecaptchaVerifierModal } from 'expo-firebase-recaptcha';
-import { GeoPoint, doc, setDoc } from 'firebase/firestore';
+import { addDoc, collection, doc, getDoc } from 'firebase/firestore';
+import { getAuth } from 'firebase/auth';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import * as Location from "expo-location";
+
 
 interface UserData {
   name: string;
   phone: string;
-  statut:string[];
-  password:string;
-  location: GeoPoint[];
+  password: string;
+  statut: string;
 }
 
+const OtpAuthConnect = ({ route, navigation }: any) => {
 
-
-
-const OtpSignUp = ({ route, navigation }: any) => {
-
-  const { phoneNumber, name, password } = route.params;
-
-  const [origin, setOrigin] = useState<any>({ latitude: 4.094354, longitude: 9.7393663,});
-
-  const initialValue: UserData = {
-    name: name,
-    phone: phoneNumber,
-    statut:["client", "chauffeur"],  
-    password:password,
-    location: origin,
-   }
-   const [localisation] = useState<UserData>(initialValue);
-
-   useEffect(() => {
-    let intervalId: NodeJS.Timeout;
-
-    const getLocationPermission = async () => {
-        let { status } = await Location.requestForegroundPermissionsAsync();
-        if (status !== "granted") {
-            alert("Permission refusée");
-            return;
-        }
-
-        const updateLocation = async () => {
-            let location = await Location.getCurrentPositionAsync({});
-            const current = {
-                latitude: location.coords.latitude,
-                longitude: location.coords.longitude,
-            };
-            setOrigin(current);
-        };
-
-        updateLocation();
-        intervalId = setInterval(updateLocation, 3000);
-    };
-
-    getLocationPermission();
-
-    return () => clearInterval(intervalId);
-}, []);
-
+  const { phoneNumber } = route.params;
   const [codeVerification, setCodeVerification] = useState('');
   const [verificationId, setVerificationId] = useState("")
-  const [utilisateur, setUtilisateur] = useState<UserData>(initialValue);
   const [renvoyer, setRenvoyer] = useState(true);
-  const [secondes, setSecondes] = useState(45);// décomptage de 10 sec par default
-  const [loading, setLoading] = useState(false); // Nouvel état pour le chargement
-  const recaptchaVerifier = useRef(null)
-  const inputRef = useRef([]);
-
- 
+  const [secondes, setSecondes] = useState(10);// décomptage de 10 sec par default
+  const [loading, setLoading] = useState(false); 
+  const recaptchaVerifier = useRef(null);
 
 
   useEffect(() => {
     const sentVerificationCode = async () => {
       try {
         const phoneProvider = new PhoneAuthProvider(auth)
-        const id = await phoneProvider.verifyPhoneNumber(`${utilisateur.phone}`, recaptchaVerifier.current!)
+        const id = await phoneProvider.verifyPhoneNumber(`${phoneNumber}`, recaptchaVerifier.current!)
         setVerificationId(id)
         Alert.alert("Alerte", "Un code vous a été envoyé par SMS!");
 
@@ -91,7 +44,7 @@ const OtpSignUp = ({ route, navigation }: any) => {
       }
     }
     sentVerificationCode();
-    }, [utilisateur.phone])
+  }, [phoneNumber])
 
   const resendCode = async () => {
 
@@ -99,7 +52,7 @@ const OtpSignUp = ({ route, navigation }: any) => {
       setSecondes(30);
       setRenvoyer(true);
       const phoneProvider = new PhoneAuthProvider(auth);
-      const id = await phoneProvider.verifyPhoneNumber(`${utilisateur.phone}`, recaptchaVerifier.current!)
+      const id = await phoneProvider.verifyPhoneNumber(`${phoneNumber}`, recaptchaVerifier.current!)
       setVerificationId(id)
       Alert.alert("Alerte", "Un nouveau code vous a été renvoyé!");
     } catch (error: any) {
@@ -119,22 +72,27 @@ const OtpSignUp = ({ route, navigation }: any) => {
       if (verificationId) {
         const credential = PhoneAuthProvider.credential(verificationId, codeVerification);
         await signInWithCredential(auth, credential);
-        
         const user = auth.currentUser;
+
         if (user) {
-        await setDoc(doc(firestore, "users", user.uid), {
-          phone: utilisateur.phone,
-          name: utilisateur.name,
-          password: utilisateur.password,
-          statut: utilisateur.statut[0],
-          location: localisation.location,
-        });
-        await AsyncStorage.setItem('userLoggedIn', 'true');
-        navigation.navigate('Home');
-      } else {
-        throw new Error("Utilisateur non trouver après connexion")
+          const userDoc = await getDoc(doc(firestore, "users", user.uid));
+          if (userDoc.exists()) {
+            const userData = userDoc.data() as UserData;
+            await AsyncStorage.setItem('userLoggedIn', 'true');
+            console.log('succes', userData.statut)
+            if (userData.statut ==="chauffeur") {
+             console.log("user data",userData.statut)
+             navigation.navigate("HomeChauffeur");
+            } else if (userData.statut) {
+              navigation.navigate("Home");
+            }
+          } else {
+            Alert.alert("Erreur", "Aucune donnée utilisateur trouvée.");
+          }
+        } else {
+          throw new Error("Utilisateur non trouvé après connexion");
+        }
       }
-    }
     } catch (error: any) {
       Alert.alert("Une erreur lors de la confirmation", error.message);
     } finally {
@@ -174,6 +132,7 @@ const OtpSignUp = ({ route, navigation }: any) => {
         <FirebaseRecaptchaVerifierModal
           ref={recaptchaVerifier}
           firebaseConfig={app.options}
+          attemptInvisibleVerification={false}
         />
         <Back />
         <Text style={{ fontSize: 25, marginTop: '5%' }}>Un code à six chiffres vous a été envoyé au : {phoneNumber}</Text>
@@ -257,4 +216,4 @@ const styles = StyleSheet.create({
 
 });
 
-export default OtpSignUp;
+export default OtpAuthConnect;
